+++ packages/_example/src/routes/+layout.svelte
@@ -1,10 +1,10 @@
 <script lang="ts">
 
-	import { Seo } from '@svaio/meta/svelte'
 	import { Head } from '@svaio/pwa/svelte'
-	import { Utils } from '@svaio/unocss/svelte'
 
 	import Header from './Header.svelte'
+	import { Seo } from '../../../tools/meta/svelte'
+	import { Utils } from '../../../tools/unocss/svelte'
 	import '../app.css'
 
 	// import * as m from '$lib/paraglide/messages.js'