+++ packages/tools/media/src/create/const.d.ts
@@ -1,0 +1,2 @@
+export declare const GLOBAL_CONST_ID = "__SVAIO_CREATE_MEDIA_PATHS__";
+export declare const PLUGIN_ID = "svaio-media-create";