+++ packages/tools/media/src/create/const.ts
@@ -1,2 +1,2 @@
-export const GLOBAL_CONST_ID = '__SVAIO_CREATE_MEDIA_PATHS__' as const
+export const GLOBAL_CONST_ID = '__SVAIO_CREATE_MEDIA_PATHS__'
 export const PLUGIN_ID = 'svaio-media-create'