+++ packages/tools/media/svelte/picture.svelte
@@ -1,0 +1,14 @@
+<script lang="ts">
+
+	import { HTMLImgAttributes } from 'svelte/elements'
+
+	let {
+		src, ...rest
+	}: HTMLImgAttributes & { src: string } = $props()
+
+</script>
+
+<enhanced:img
+	src={src}
+	{...rest}
+/>