+++ packages/tools/media/svelte/utils.ts
@@ -1,5 +1,6 @@
 /* eslint-disable jsdoc/check-tag-names */
-import { GLOBAL_CONST_ID } from './const'
+
+import type { GLOBAL_CONST_ID } from '../src/create/const'
 
 /**
  * Get Created Media paths
@@ -9,8 +10,9 @@ import { GLOBAL_CONST_ID } from './const'
  */
 export const getCreateMediaPaths = (): Record<string, string> => {
 
+	const ID: typeof GLOBAL_CONST_ID = '__SVAIO_CREATE_MEDIA_PATHS__'
 	// @ts-ignore
-	const obj = globalThis?.[GLOBAL_CONST_ID]
+	const obj = globalThis?.[ID]
 
 	if ( typeof obj === 'object' && obj !== null )
 		return obj as Record<string, string>