+++ packages/tools/pwa/build.config.ts
@@ -2,4 +2,9 @@
 import { config }            from '@svaio/repo-config/unbuild'
 import { defineBuildConfig } from 'unbuild'
 
-export default defineBuildConfig( [ { ...config } ] )
+export default defineBuildConfig( [
+	{
+		...config,
+		failOnWarn : false,
+	},
+] )