+++ packages/_example/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @svaio\_/example
 
+## 0.0.5
+
+### Patch Changes
+
+- fix svelte types
+
 ## 0.0.3
 
 ### Patch Changes