+++ packages/core/CHANGELOG.md
@@ -1,5 +1,16 @@
 # svaio
 
+## 0.0.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @svaio/sitemap@0.0.5
+  - @svaio/unocss@0.0.5
+  - @svaio/media@0.0.5
+  - @svaio/meta@0.0.5
+  - @svaio/pwa@0.0.5
+
 ## 0.0.4
 
 ### Patch Changes