+++ packages/tools/unocss/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @svaio/unocss
 
+## 0.0.5
+
+### Patch Changes
+
+- fix svelte types
+
 ## 0.0.4
 
 ### Patch Changes